initial work to add functionality to extend app definition used by cards navigation
