+++ src/plugins/management/public/types.ts
@@ -10,7 +10,7 @@ import { Observable } from 'rxjs';
 import { ScopedHistory, Capabilities } from '@kbn/core/public';
 import type { LocatorPublic } from '@kbn/share-plugin/common';
 import { ChromeBreadcrumb, CoreTheme } from '@kbn/core/public';
-import type { AppId } from '@kbn/management-cards-navigation';
+import type { CardsNavigationComponentProps } from '@kbn/management-cards-navigation';
 import { AppNavLinkStatus } from '@kbn/core/public';
 import { ManagementSection, RegisterManagementSectionArgs } from './utils';
 import type { ManagementAppLocatorParams } from '../common/locator';
@@ -82,9 +82,9 @@ export interface CreateManagementItemArgs {
   redirectFrom?: string; // redirects from an old app id to the current app id
 }
 
-export interface NavigationCardsSubject {
+export interface NavigationCardsSubject
+  extends Pick<CardsNavigationComponentProps, 'hideLinksTo' | 'extendAppDefinitons'> {
   enabled: boolean;
-  hideLinksTo?: AppId[];
 }
 
 export interface AppDependencies {
